@@ -1,8 +1,8 @@
 {
   "scripts": {
-    "start": "react-native start",
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
+    "start": "expo start --dev-client",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web",
     "eject": "expo eject",
     "test": "jest -u",
@@ -15,8 +15,9 @@
     "@react-native-community/datetimepicker": "3.5.2",
     "@react-native-community/hooks": "2.6.0",
     "@react-native-community/masked-view": "0.1.10",
-    "@react-native-community/netinfo": "6.0.0",
-    "@react-navigation/native": "5.9.4",
+    "@react-native-community/netinfo": "6.0.2",
+    "@react-navigation/native": "^6.0.6",
+    "@react-navigation/native-stack": "^6.2.5",
     "@types/geojson": "7946.0.8",
     "@types/react-native-autocomplete-input": "^4.0.2",
     "@types/react-native-simple-radio-button": "2.7.1",
@@ -24,16 +25,16 @@
     "apollo-boost": "0.4.9",
     "axios": "0.21.1",
     "buffer": "6.0.3",
-    "expo": "42.0.4",
-    "expo-file-system": "~11.1.3",
-    "expo-image-picker": "^10.2.3",
-    "expo-linear-gradient": "~9.2.0",
-    "expo-location": "~12.1.2",
-    "expo-permissions": "~12.1.1",
-    "expo-splash-screen": "~0.11.2",
-    "expo-status-bar": "~1.0.4",
-    "expo-task-manager": "~9.2.2",
-    "expo-web-browser": "~9.2.0",
+    "expo": "^43.0.0",
+    "expo-file-system": "~13.0.3",
+    "expo-image-picker": "~11.0.3",
+    "expo-linear-gradient": "~10.0.3",
+    "expo-location": "~13.0.4",
+    "expo-permissions": "~13.0.3",
+    "expo-splash-screen": "~0.13.5",
+    "expo-status-bar": "~1.1.0",
+    "expo-task-manager": "~10.0.3",
+    "expo-web-browser": "~10.0.3",
     "form-data": "4.0.0",
     "geojson": "0.5.0",
     "graphql": "14.6.0",
@@ -42,30 +43,29 @@
     "lodash": "4.17.21",
     "mathjs": "^9.4.4",
     "moment": "2.29.1",
-    "react": "16.13.1",
+    "react": "17.0.1",
     "react-apollo": "3.1.5",
     "react-devtools": "4.13.5",
-    "react-dom": "16.13.1",
+    "react-dom": "17.0.1",
     "react-hook-form": "7.10.1",
     "react-i18next": "11.11.1",
     "react-moment": "1.1.1",
-    "react-native": "~0.63.4",
+    "react-native": "0.64.3",
     "react-native-autocomplete-input": "5.0.2",
     "react-native-elements": "3.4.2",
     "react-native-gesture-handler": "~1.10.2",
     "react-native-logs": "3.0.4",
-    "react-native-maps": "0.28.0",
+    "react-native-maps": "^0.29.3",
     "react-native-modal": "12.0.2",
     "react-native-reanimated": "~2.2.0",
-    "react-native-safe-area-context": "3.2.0",
-    "react-native-screens": "~3.4.0",
+    "react-native-safe-area-context": "3.3.2",
+    "react-native-screens": "~3.8.0",
     "react-native-shadow-2": "^3.0.0",
     "react-native-simple-radio-button": "2.7.4",
     "react-native-svg": "12.1.1",
     "react-native-typescript-geojson": "0.0.4",
-    "react-native-unimodules": "0.14.10",
     "react-native-uuid": "2.0.1",
-    "react-native-web": "~0.13.12",
+    "react-native-web": "0.17.1",
     "react-navigation": "4.4.4",
     "react-navigation-stack": "2.10.4",
     "react-redux": "7.2.4",
@@ -74,19 +74,20 @@
     "redux-thunk": "2.3.0"
   },
   "devDependencies": {
-    "@babel/core": "~7.9.0",
+    "@babel/core": "^7.12.9",
     "@types/jest": "26.0.24",
     "@types/lodash": "4.14.171",
-    "@types/react": "~16.9.35",
-    "@types/react-native": "~0.63.2",
+    "@types/react": "~17.0.21",
+    "@types/react-native": "~0.64.12",
     "@types/react-redux": "7.1.18",
     "@types/react-test-renderer": "17.0.1",
     "babel-eslint": "10.1.0",
-    "babel-preset-expo": "8.3.0",
+    "babel-preset-expo": "8.5.1",
     "eslint": "7.30.0",
     "eslint-plugin-react": "7.24.0",
-    "jest-expo": "^42.0.0",
-    "typescript": "~4.0.0"
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "jest-expo": "^43.0.0",
+    "typescript": "~4.3.5"
   },
   "jest": {
     "preset": "jest-expo",
