@@ -7,7 +7,8 @@ module.exports = {
     },
     'extends': [
         'eslint:recommended',
-        'plugin:react/recommended'
+        'plugin:react/recommended',
+        'plugin:react-hooks/recommended'
     ],
     'globals': {
         'Atomics': 'readonly',
@@ -70,25 +71,25 @@ module.exports = {
     "settings": {
         "react": {
             "createClass": "createReactClass", // Regex for Component Factory to use,
-                                              // default to "createReactClass"
+            // default to "createReactClass"
             "pragma": "React",  // Pragma to use, default to "React"
             "fragment": "Fragment",  // Fragment to use (may be a property of <pragma>), default to "Fragment"
             "version": "detect", // React version. "detect" automatically picks the version you have installed.
-                                // You can also use `16.0`, `16.3`, etc, if you want to override the detected value.
-                                // default to latest and warns if missing
-                                // It will default to "detect" in the future
+            // You can also use `16.0`, `16.3`, etc, if you want to override the detected value.
+            // default to latest and warns if missing
+            // It will default to "detect" in the future
             "flowVersion": "0.53" // Flow version
         },
         "propWrapperFunctions": [
-              // The names of any function used to wrap propTypes, e.g. `forbidExtraProps`. If this isn't set, any propTypes wrapped in a function will be skipped.
-              "forbidExtraProps",
-              {"property": "freeze", "object": "Object"},
-              {"property": "myFavoriteWrapper"}
+            // The names of any function used to wrap propTypes, e.g. `forbidExtraProps`. If this isn't set, any propTypes wrapped in a function will be skipped.
+            "forbidExtraProps",
+            { "property": "freeze", "object": "Object" },
+            { "property": "myFavoriteWrapper" }
         ],
-          "linkComponents": [
+        "linkComponents": [
             // Components used as alternatives to <a> for linking, eg. <Link to={ url } />
             "Hyperlink",
-            {"name": "Link", "linkAttribute": "to"}
+            { "name": "Link", "linkAttribute": "to" }
         ]
     }
 }
