@@ -1,44 +1,37 @@
-import React, { Component } from 'react'
-import { createAppContainer } from 'react-navigation'
-import { createStackNavigator } from 'react-navigation-stack'
-import { withTranslation, WithTranslation } from 'react-i18next'
+import React from 'react'
+import { NavigationContainer } from '@react-navigation/native'
+import { createNativeStackNavigator, NativeStackNavigationOptions } from '@react-navigation/native-stack'
+import { withTranslation } from 'react-i18next'
 import HomeScreen from '../screens/HomeScreen'
 import MapScreen from '../screens/MapScreen'
 import ObservationScreen from '../screens/ObservationScreen'
 import ObservationEventScreen from '../screens/ObservationEventScreen'
 import LoginScreen from '../screens/LoginScreen'
 import EditObservationEventScreen from '../screens/EditObservationEventScreen'
-import InfoScreen from '../screens/InfoScreen'
+import NavBarComponent from '../components/general/NavBarComponent'
 
-const Navigator = createStackNavigator(
-  {
-    Login: LoginScreen,
-    Home: HomeScreen,
-    Map: MapScreen,
-    Observation: ObservationScreen,
-    ObservationEvent: ObservationEventScreen,
-    EditObservationEvent: EditObservationEventScreen,
-    Info: InfoScreen
-  },
-  {
-    initialRouteName: 'Login'
-  }
-)
+const Stack = createNativeStackNavigator()
 
-const AppContainer = createAppContainer(Navigator)
-
-class AppNavigator extends Component<WithTranslation> {
-  render() {
-    const { t, i18n } = this.props
-
-    return (
-      <AppContainer
-        screenProps={{
-          t,
-          i18n
-        }}
-      />
-    )
+const navigationOptions: NativeStackNavigationOptions = {
+  header: function Header ({ navigation, route }) {
+    return (<NavBarComponent navigation={navigation} route={route} />)
   }
 }
-export default withTranslation()(AppNavigator)
+
+const Navigator = () => {
+
+  return (
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName='login'>
+        <Stack.Screen name='login' component={LoginScreen} options={navigationOptions} />
+        <Stack.Screen name='home' component={HomeScreen} options={navigationOptions} />
+        <Stack.Screen name='map' component={MapScreen} options={navigationOptions} />
+        <Stack.Screen name='observation' component={ObservationScreen} options={navigationOptions} />
+        <Stack.Screen name='overview' component={ObservationEventScreen} options={navigationOptions} />
+        <Stack.Screen name='document' component={EditObservationEventScreen} options={navigationOptions} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  )
+}
+
+export default withTranslation()(Navigator)
